BREAKING(get rid of inner component)
