+++ addon-test-support/is-visible.js
@@ -1,14 +1,13 @@
 import { find } from 'ember-native-dom-helpers';
 
 export function isVisible(selector, contextEl) {
-  const inner = typeof(selector) === 'string' ? find(`${selector} > .inner`, contextEl)
-    : find('.inner', selector);
+  const attachment = typeof(selector) === 'string' ? find(selector, contextEl) : selector;
 
-  if (!inner) {
+  if (!attachment) {
     const msg = typeof(selector) === 'string' ? selector : selector.innerHTML;
 
     throw `Could not locate attachment from selector: ${msg}`;
   }
 
-  return inner.style.display !== 'none';
+  return attachment.style.display !== 'none';
 }
