+++ addon/styles/addon.scss
@@ -6,14 +6,12 @@ $origins: bottom, top, right, left;
 .ember-attacher-popper {
   min-height: 10px;
 
-  > .inner {
-    max-width: 400px;
-    perspective: 800px;
-    z-index: 9999;
-  }
+  max-width: 400px;
+  perspective: 800px;
+  z-index: 9999;
 
   @each $position in $positions {
-    &[x-placement=#{$position}] > .inner {
+    &[x-placement=#{$position}] {
       > div[x-arrow] {
         @include arrow($position, 10px);
       }
@@ -86,7 +84,7 @@ $origins: bottom, top, right, left;
   }
 }
 
-.ember-attacher-tooltip > .inner {
+.ember-attacher-tooltip {
   position: relative;
   color: white;
   border-radius: 4px;
@@ -134,7 +132,7 @@ $origins: bottom, top, right, left;
   }
 }
 
-.ember-attacher-light-theme > .inner {
+.ember-attacher-light-theme {
   color: #203d5d;
   box-shadow: 0 0 16px -4px rgba(0,20,40,0.2), 0 0 80px -10px rgba(0,20,40,0.3);
   background-color: #f3f6f9;
@@ -151,7 +149,7 @@ $origins: bottom, top, right, left;
 }
 
 @media (max-width: 500px) {
-  .ember-attacher-tooltip > .inner, .ember-attacher-popover > .inner {
+  .ember-attacher-tooltip , .ember-attacher-popover {
     max-width: 94%;
     max-width: calc(100% - (2rem + 4px));
   }
