--- addon/templates/components/ember-attacher-inner.hbs
@@ -1,10 +1,0 @@
-{{#if _shouldRender}}
-  {{yield (hash hide=(action 'hide'))}}
-
-  {{#if arrow}}
-    <div x-arrow></div>
-  {{/if}}
-  {{#if (eq animation 'fill')}}
-    <div x-circle style="{{circleTransitionDuration}}"></div>
-  {{/if}}
-{{/if}}
