--- addon/templates/components/ember-attacher.hbs
@@ -1,32 +1,0 @@
-{{#ember-popper ariaRole=ariaRole
-                class=class
-                eventsEnabled=false
-                id=id
-                modifiers=_modifiers
-                onFoundTarget=(action 'onFoundTarget')
-                placement=placement
-                popperContainer=popperContainer
-                renderInPlace=renderInPlace
-                target=target as |emberPopper|}}
-  {{#ember-attacher-inner animation=animation
-                          arrow=arrow
-                          class=(concat 'inner ' innerClass)
-                          disableEventListeners=emberPopper.disableEventListeners
-                          enableEventListeners=emberPopper.enableEventListeners
-                          hideDelay=hideDelay
-                          hideDuration=hideDuration
-                          hideOn=hideOn
-                          interactive=interactive
-                          isOffset=isOffset
-                          isShown=isShown
-                          lazyRender=lazyRender
-                          onChange=onChange
-                          update=emberPopper.update
-                          showDelay=showDelay
-                          showDuration=showDuration
-                          showOn=showOn
-                          target=_target as |inner|}}
-    {{yield (hash emberPopper=emberPopper
-                  hide=inner.hide)}}
-  {{/ember-attacher-inner}}
-{{/ember-popper}}
