--- app/components/ember-attacher-inner.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-attacher/components/ember-attacher-inner';
