--- app/components/ember-attacher.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-attacher/components/ember-attacher';
